[FEATURE] Add an AbstractHtmlProcessor class
This class will be the base class for other classes that can remove
invisible nodes, convert CSS to non-CSS attributes etc.

The folder structure will be cleaned up once the Emogrifier class has
been renamed to CssInliner and moved from the Pelago\ namespace to the
Pelage\Emogrifier namespace (which will happen in Emogrifier 4.0.0).