[BUGFIX] Descendent attribute selectors
Changed CSS-to-XPath translation matcher for descendant combinator to exclude
only spaces within `[...]`, i.e. when not followed by a closing `]` unless there
is an intervening opening `[` (rather than than excluding any space where the
remaining part of the expression ends with a closing `]`).  This matches spaces
in the original CSS selector but not spaces in already-translated XPath
predicates, and allows attribute selectors to be used in descendants.

Added PHPUnit tests for some attribute selectors as descendants, particularly
where the translated XPath predicate will contain a space.  Also added tests for
the converse (attribute selectors with descendants), and comments for relevant
tests that should be added but would currently fail due to other issues.

This should fix Issue
[#381](https://github.com/MyIntervals/emogrifier/issues/381) and at least
partially address Issue
[#443](https://github.com/MyIntervals/emogrifier/issues/443).

For reference, the original space-exclusion regex was added in
[#327](https://github.com/MyIntervals/emogrifier/pull/327) to avoid matching
spaces in selector attributes.  The purpose changed in
[#458](https://github.com/MyIntervals/emogrifier/pull/458), when the
CSS-to-XPath translation rules were reordered, to avoid matching spaces in XPath
predicates, though the requirement not to match spaces within `[...]` was
unchanged.