[CLEANUP] `splitCssAndMediaQuery`: retain ordering
Return from `splitCssAndMediaQuery` an array of split CSS parts for different
media queries, along with the parts of intervening CSS not within a media query,
in order.

The ordering will need to be retained for implementing #280.

Currently in `process` the return value is converted back to `$cssParts` as it
was before, as a temporary measure, pending further refactoring, of
`parseCssRules` and `copyCssWithMediaToStyleNode`.