[CLEANUP] Tighten tests using `assertNotContains`
In tests using `assertNotContains`, assert only that a minimal string sufficient
to indicate a test failure is not present.

Changes for #280 will inevitably mean that during processing, due to the way
that some components of the CSS are `trim`med, some unnecessary whitespace is
removed from the CSS for media queries added to the style element.  Thus
asserting that a complete media query rule is not present in the result may
allow a test to pass that should fail.  Asserting that just, e.g., "@media" or
"@media tv" is not present in the result gives a more stringent test.

For completeness, needles used with `assertNotContains` in other tests have also
been similarly shortened.  Some of these tests would have allowed invalid
passes, e.g. with a semicolon at the end of a style attribute that was asserted
not to be present.