[CLEANUP] Allow whitespace removal in @media tests (#521)
Added test assertion methods similar to `assertContains` and `assertNotContains`
but which also allow for removal of some unnecessary whitespace from the CSS,
without having to make the test data less humanly readable or change the test
inputs.

Changes for #280 will inevitably mean that during processing, due to the way
that some components of the CSS are `trim`med, some unnecessary whitespace is
removed from the CSS for media queries added to the style element.  Relevant
tests have been changed to used the new assertion methods.