[CLEANUP] Refactor `parseCssRules`: parse all CSS
Refactored `parseCssRules` so that it can now parse an entire CSS string,
including @media rules.  It now returns two arrays, in an array with the
following keys:
- "inlineable" => The inlineable rules that would have been previously returned;
- "uninlineable" => The uninlineable rules, currently those from within @media
  rules.

The original line number information from the CSS is available in both arrays,
so the original ordering of all the rules (that will be required for #280) can
be determined.

A temporary method `combineCssMediaRules` has been added to recombine the @media
rules back into a string of CSS as expected by `copyCssWithMediaToStyleNode`,
pending refactoring of that method to instead accept a parsed rule array as its
parameter.  This does mean that, for now, the rules within each @media rule are
parsed again.