[CLEANUP] Symfony-compatible specificity tests
Modified selector specificity tests using universal selectors and combinators in
the `:not` argument (for ":not alone") to use a single type instead here, with
the other, equally-specific selector in each test also gaining a type.  This is
because the Symfony CssSelector component does not support combinators in the
:not argument.

Relates to #540.