[FEATURE] Add a renderBodyContent() method
For `CssInliner` and `AbstractHtmlProcessor`, this method is part of the
public API. For `Emogrifier`, it is internal.

This is part of #554.

Also autoformat the changed classes, fix a typo in a test name, and remove
the check for the existence of the BODY element from `getBodyElement()`
(as this always is ensured and hence will never happen).