[BUGFIX] Allow `:not(:behavioural-pseudo-class)` (#697)
This was causing an error because `removeUnmatchablePseudoComponents()` was
converting the selector to `:not()`, which is invalid.  The intention would have
been to convert it to `:not(*)`, but this is not correct either.  If a
behavioural pseudo-class is used within the `:not` argument, then the entire
`:not(…)` component should be stripped in order to test whether the selector
could match any DOM elements.