[FEATURE] Keep `display:none` elements with `-emogrifier-keep` (#737)
The additional part of the XPath expression was obtained by running
`:not(.-emogrifier-keep)` through `CssSelectorConverter::toXPath()`.  It tests
for attribute presence before checking the value, presumably as an optimization
to avoid unnecessary processing of a null/empty value.  On the presumed basis
that this is a good idea, a similar check has been added for the `style`
attribute.

Fixes #252.