[TASK] Test :nth-last-child (#750)
Also updated the regex for supported pseudo-classes (`PSEUDO_CLASS_MATCHER`) so
that those supported are explicitly specified (rather than ‘anything ending with
`-child` or anything ending with `-type(`’), without acually affecting the valid
pseudo-class names it will match.  Except in the `Emogrifier` class – this
doesn’t support `:nth-last-child` so that has been removed (to the effect that
rules with it will now be copied to the `<style>` element rather than attempted
to be inlined).

And added `:nth-last-child` to the list of supported pseudo-classes in the
README.  (Also, for consistency, added `()` to the other `:nth…` pseudo-classes
in that section.)

Part of #747.