[TASK] Test :any-link
Added a “broken: …” comment placeholder to indicate that support for `:any-link`
with the Symfony CssSelector component has been checked and found to be lacking,
suggesting tests to be added if and when it is supported in future.

Added test to confirm rules with `:any-link` are nonetheless copied to a
`<style>` element.  Also added similar tests for `:nth-last-child`,
`:nth-last-of-type` and `:only-child` with the `Emogrifier` class (these are
fully supported only with the `CssInliner` class).

In the README, added `:any-link` to a list of unsupported static pseudo-classes
that are copied to a `<style>` element rather than being inlined.

Part of #748.