[TASK] Test :only-of-type
Added a “broken: …” comment placeholder to indicate that support for
`:only-of-type` with the Symfony CssSelector component has been checked and
found to be broken (see https://github.com/symfony/symfony/issues/33773),
suggesting tests to be added when it is fixed.

(It does not make sense to support a feature that is broken the way it is – for
email clients that support `:only-of-type`, the desired result will currently
only be achieved if rules involving it are copied to a `<style>` element, rather
than being only partially inlined.)

Added test to confirm rules with `:only-of-type` are copied to a `<style>`
element.

In the README, added `:only-of-type` to the list of unsupported static
pseudo-classes that are copied to a `<style>` element rather than being inlined.

Part of #748.