[TASK] Test :optional and :required
Added a “broken: …” comment placeholder to indicate that support for `:required`
and `:optional` with the Symfony CssSelector component has been checked and
found to be lacking, suggesting tests to be added if and when it is supported in
future.

Added tests to confirm rules with `:required` or `:optional` are nonetheless
copied to a `<style>` element.

In the README, added `:required` and `:optional` to the list of unsupported
static pseudo-classes that are copied to a `<style>` element rather than being
inlined.

Closes #748.