[FEATURE] Composer script for PHPUnit with --stop-on-failure (#782)
During development, if something is a bit awry, it’s likely that a whole bunch
of tests will fail, producing reams of output about all the failures.  In this
situation, it’s preferable to be able to run the tests so that they will stop on
the first failure.

The name of the script is deliberaly short – it needs to be manually typed (at
least once), often in frustration!