[BUGFIX] partially satisfy psalm, updating baseline
This commit resolves the PossiblyInvalidArgument issues identified in
tests/Unit/HtmlProcessor/HtmlPrunerTest.php by using a pair of
psalm-specific docblock tags (`@psalm-return` and `@psalm-var`).

The `@psalm-return` and `@psalm-var` tags are used here instead of
altering the existing `@return` tag and adding a standard `@var` tag
as some CI tools & services do not support psalm's array shapes.

The `@psalm-var` tag is paired with each `@psalm-return` tag to force
psalm to agree on the return type; The actual return types use a set of
specific string keys rather than ints, but this method of pairing
psalm-specific tags allows the existing meaning of the method bodies
to be retained without excessive changes.

regarding: #537