[FEATURE] Add support for PHP 7.4
Also only run the CI tasks that do not depend on the PHP version
(i.e., everything except for the PHP linting and the unit tests)
on PHP 7.3 only to avoid problems with PHP 7.4.

Also ignore the platform reqs for PHP 7.3 and 7.4 during composer update
for the time being.

Fixes #819