[TASK] Normalize the DOCTYPE declaration
Ensure that the DOCTYPE declaration consists of uppercase `DOCTYPE` and
lowercase root element name (`html`).

This is done when the `DOMDocument` is created from an HTML source.  Once the
`DOMDocument` has been created, the `DOMDocumentType` cannot be changed, so the
document type declaration must be manipulated (if necessary) in the HTML
beforehand.  (Since only HTML documents are supported, the declaration is only
normalized when the root element name is HTML, in whatever case - the precise
specification for any element name involves lists of various Unicode character
ranges which it would be superfluous to allow for and try to match.  PHP's
`DOMDocument`/`libxml` itself will output the `DOCTYPE` keyword in uppercase in
any case.)

This normalization is consistent with the relevant part of the
[polyglot markup specification](
  https://dev.w3.org/html5/html-polyglot/html-polyglot.html#doctype
).
 While polyglot markup is primarily intended for serialization of HTML as XML
(we don't actually support outputting as XHTML), is also recommended for maximum
interoperability and robustness when rendering HTML.

This also makes the output consistent with that of `Masterminds/html5-php` and
would eliminate the need to change associated tests specifically for #831.

Closes #858.