[BUGFIX] Avoid exception with :last-of-type etc.
This simply broadens the range of exceptions thrown by `symfony/css-selector`
(for unsupported selector expressions) which will be caught and discarded in
non-debug mode.

It does not make anything magically work which didn't previously; however, it
does allow the rest of the CSS to achieve full functionality, with only the
affected CSS rule and selector suffering.

The tests and changes to the README clarify and correct-as-documented the
current behaviour for these edge cases, which could be improved upon in future.

These are currently rarely-used CSS pseudo-classes; their level of support in
mainstream browsers has only
[recently reached 95% of audiences](https://caniuse.com/#search=last-of-type).
 Their level of support in email clients is likely to be drasitcally lower per
se of some time to come, but may be something Emogrifier can help with moving
forwards, as their use becomes more widespread...

Closes #872.  New issue #876 for follow-up.