[TASK] Rename the PHPCS configuration file (#891)
There are two possible file names for the PHP_CodeSniffer configuration
file:

- `phpcs.xml.dist`
- `phpcs.xml`

The intention of the first file name is: "This is a template. Please copy
it, modify it and then save it under the corresponding file name without
the .dist` suffix."

As our configuration file is not a template, we now use the non-template
variant.

As an additional benefit, IDEs usually recognize the `.xml` suffix
without any additional configuration.