[BUGFIX] Allow Windows to run Phive Phars
The Windows CLI is not smart enough to inspect an 'executable' file (e.g. to
look for a 'shebang') to determine what program is required to run it.  It
requires the file to have a file extension which is associated with the required
program.  Thus the Phar files have been renamed to have the `.phar` extension.