[BUGFIX] Check existing Content-Type is in <head>
If a valid `Content-Type` `<meta>` element is present, DOM conversion will
create a `<head>` element for it even without an explicit `<head>` tag in the
HTML.

However, to be valid, it must not be in the `<body>` element.  As well as with
an explicit `<body>` start tag, the `<body>` element also begins whenever a
start tag for an element which cannot be in the `<head>` is encountered.  This
is now checked.

Fixes #923.