[BUGFIX] Keep PhpStorm from indexing the tool PHARs (#968)
PHPUnit is the only tool for which we would like PhpStorm to index
the PHAR (so that it knows the PHPUnit classes exist). For all other
tools, indexing the PHARs would only slow PhpStorm down (and possibly
create warnings about classing existing multiple times).

Use Composer's PHP process for running the tool PHARs.
This allows running the PHARs without the `.phar` suffix even
on Windows.