[CLEANUP] Add type hints for stdClass object properties
Psalm supports `object{}` type hints to provide type hints for the properties
of a `stdClass` object.  Using these in the DocBlocks allows various type hints
and (redundant) casts within the code to be eliminated.

PHPCS however does not currently understand them, so some DocBlock sniffs are
locally disabled for now.