[CLEANUP] Remove a redundant CSS data cache
The cache with the key `CACHE_KEY_CSS` was storing all of the parsed CSS data
as a single entry, but would only ever be set once and never accessed.  It does
not make sense to call `inlineCss` again on the same document with the same CSS,
but in any case this method clears all caches as its first action.

This will reduce the memory footprint after `inlineCss` returns.