[TASK] getCssRegularExpression without delimiters
This will allow the method, correspondingly renamed from
`getCssNeedleRegularExpressionPattern` to `getCssRegularExpressionMatcher`, to
be used more flexibly, not necessarily for finding needles in haystacks.  It
also allows slight simplifications and tightening-up of its tests.

Update the DocBlock for the method to reflect this change and previous changes.

Also now compute the regular expression version of the CSS in the `Constraint`
constructors - this will be more efficient if the `Constraint` is re-used.