[CLEANUP] Move CssConstraint DocBlock info re c'tor there
Remove the information about the constructor for `CssConstraint` from the
DocBlock for the class, and add it to that for the constructor where it belongs.

Also further explain why the constructor is retained (see #1002).

Also change constructor visibility to `protected` (for good measure) - the class
is not intended to be instantiated in itself (and is `abstract` anyway).