[TASK] Remove irrelevant trailing zeros from test data
Remove trailing zeros from decimal numbers in CSS test data that are unrelated
to the specifics of whether numerical equivalence is maintained.  CSS parsing
may result in such numbers being re-rendered more efficiently (i.e. `1.30`
becomes `1.3` and `1.0` becomes `1`).  The changed test data is not concerned
with this; it is for confirming uninlineable at-rules as a whole are passed on
roughly equivalently.