[CLEANUP] Create @import matcher in initial CSS pass
This avoids the added complexitiy of trying to match within a regular expression
that has been generated to match the original CSS.  Now the originally-provided
plain CSS is matched instead (for `@import` which may or may not use the CSS
`url` function and may or may not have the URL within quotes).

It also avoids having to update this if a change is made to how the CSS matcher
is generated which would affect how part of an `@import` rule is matched, as was
the case with #1058.