[CLEANUP] Drop an unused naming rule from the PHPMD ruleset
We actually think that long variable names can be a good thing,
and we do not want PHPMD to stop us from using them.

In order to avoid the impression of "we need to fix our code before
we can enable the commented-out rule", the rule now has been
removed from our PHPMD ruleset altogether.