[TASK] Remove the Dependabot configuration for Composer updates
We'll update our dependencies manually if and when the need arises
(as the Composer updates were the majority of the PRs by Dependabot).

Dependabot will still update the GitHub actions.

Dependabot will still create PRs for security-relevant updates, though.

Also drop the target milestone from the configuration in order to
reduce the work needed after we have released a new major version.

Fixes #1174