[CLEANUP] Remove unnecessary `md5()` call in `CssInliner::getCssSelectorPrecedence()`
The input selector is guaranteed to be a string, so we can directly index the
cache array with it. This avoids the overhead of calculating the cryptographic
hash and the creation of a fresh string, makes the array more friendly for
manual inspection using a debugger and avoids using a function that might be
deprecated with PHP 8.4.