[CLEANUP] Add class for preg_ function error handling
The new class uses functionality that previously existed in `CssInliner`.
The `preg_replace` wrapper is extended to also support arrays as the `$pattern`
and `$replacement` parameters, and to support the `$limit` and `$count`
parameters.  A `preg_replace_callback` wrapper is also added.

With these methods in a separate class, they can be independently tested,
which is duly done.

All calls to `preg_replace` and `preg_replace_callback` are changed to use the
new class.  This resolves a number of PHPStan errors.