[TASK] Remove support for PREG_SPLIT_OFFSET_CAPTURE from Preg::split
This flag is not used anywhere in the current code base.  It causes
complications with static analysis due to its introduction of a dynamic return
type for `preg_split`, which would require a PHPStan extension to properly solve
(and which assumes PHPStan would always be used as the static analysis tool).