[CLEANUP] Remove an impossible `throw`
PHPStan is able to determine that `array_combine` when passed the results of
`array_keys` and `array_values` for the same array can never fail (in the sense
of returning `false`), because the two passed arrays will always be the same
length.

Psalm could not, which was why the return value check and `throw` was originally
put in.