[CLEANUP] Add DeclarationBlockParser class (#1311)
This is essentially a refactoring to move the `parseCssDeclarationsBlock` method
that was duplicated in both `CssInliner` and `CssToAttributeConverter` to a
separate class.

As well as eliminating duplicate code/functionality, there are other benefits:
- The method can be tested independently, which is duly done;
- The results cache is now shared;
- The functionality is readily available to other classes, e.g. for the solution
  to #1276 (as currently proposed).

PHPStan errors from the original code have been resolved as follows:
- Use `Preg::split` to wrap `preg_split` to handle unexpected errors;
- Check the return value from `preg_match` precisely with `!== 1` instead of
  `!`.

Longer-term, we should be able to use functionality from PHP-CSS-Parser for more
accurate parsing (e.g. strings containing semicolons), whereupon this class can
become a wrapper for that, but currently that functionality is not available in
a standalone class.