[CLEANUP] Remove inapplicable comment in DeclarationBlockParser::parse (#1321)
This was added during development while `Preg::split` supported
`PREG_SPLIT_OFFSET_CAPTURE` and thus could return an array of arrays, and the
code contained a check for this.  But before #1311 was completed, #1313 was
implemented to avoid the headache of a dynamic return type.  The check was
removed, but not the corresponding comment explaining it.