[CLEANUP] Simplify a data provider (#1407)
This change avoids the nested `array_map` calls and makes the
data provider shorter and easier to understand.

Also, not using `array_*` functions paves the way towards
switching to `rawr/phpunit-data-provider`, which returns
iterables instead of array (which means that we cannot use
array functions on the results provided by that library's
methods).

Part of #1317