[BUGFIX] Allow single character CSS properties (#1506)
There are no real-world examples, but the specification does allow for this:

https://drafts.csswg.org/css-syntax-3/#ident-token-diagram

This potential issue was introduced by a tiny mistake in #1500,
which already has a changelog entry.
As there has not been a release since,
no changelog entry is needed for this change.

Also add another possessive quantifier to the regex to improve performance.